LMN-420 Remove custom semantics
There was an issue where this was announcing various different slider options with percentage/value based content. This provided a confusing user behaviour.
The current behaviour isn't perfect, but we'd need to build a completely custom slider to solve this. We can create an issue for the slider to improve. the API/accessibility support