lumen/LMN-420 update Range slider to support custom thumb with label (#1745)
* lumen/LMN-420_update_slider_to_provide_custom_thumb

* Updated snapshots for 'rtl'

* Updated snapshots for 'default'

* LMN-420 Remove custom semantics

There was an issue where this was announcing various different slider options with percentage/value based content. This provided a confusing user behaviour.
The current behaviour isn't perfect, but we'd need to build a completely custom slider to solve this. We can create an issue for the slider to improve. the API/accessibility support

* added missing liecnes

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>
Co-authored-by: Maria Neumayer <maria.neumayer@skyscanner.net>